added chart values to stat

+++ app/src/main/java/dev/gbenga/inaread/data/model/MeterMonthlyStat.kt
@@ -2,11 +2,12 @@ package dev.gbenga.inaread.data.model
 
 import dev.gbenga.inaread.ui.home.DefaultInaTextIcon
 import dev.gbenga.inaread.ui.home.InaTextIcon
+import java.math.BigDecimal
 
 private val defaultIcon = DefaultInaTextIcon()
 data class MeterMonthlyStat(
     val lifeTimeReading: InaTextIcon = defaultIcon,
     val monthlyStat: InaTextIcon = defaultIcon,
     val costStat: InaTextIcon = defaultIcon,
-    val chartData: List<Float> = emptyList()
+    val chartData: List<BigDecimal> = emptyList()
 )
